Add Coptic layout.
Coptic layout based on the soft keyboard available at
https://coptic-dictionary.org/ Some additional diacritics
are added to punctuation keys for clear unicode transcription.

Number keys are derived from a us layout. It should be usable
with either an ansi or iso physical layout.

The layout is entered under Egypt's country code, which reflects
its historical range, although it might make more sense to group
it either with other ancient languages or to list it under a 'cop'
group referring to the ethnic and religious diaspora.