rules: rename the files in sort order
Having a magic order in a variable in the Makefile (and having it conditional
on the compat rules) is messy and makes life harder for everyone.

Since we cannot change the order of the files anyway, let's rename them so we
can rely on standard sorting methods to give us the right order. This
makes filtering significantly easier, we can just append the compat scripts
when needed and rely on our (modified) merge.py script to sort by basename.

Plus, it makes it more obvious to new contributors that these files do indeed
have a fixed order that must not be changed.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>