man: swap __ for @ in the sed substitutions
meson cannot replace __foo__ easily, it's built to substitute only @foo@. So
as prep-work let's switch to that notation but add an extra sed to swap that
back to the __foo__ notation for autotools. That one can't switch away from
that either because it relies on MAN_SUBST which is filled in by
xorg-macros.m4.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>