gitlab CI: use ci-fairy to check merge requests
The fdo ci-templates provide a tool called ci-fairy which can check
merge requests as part of a pipeline. In this case we check for the checkbox
set to allow collaboration (i.e. the one that allows a maintainer to rebase a
merge request). Where it is missing, the CI job fails (though it's not fatal).

ci-fairy uses FDO_UPSTREAM_REPO to get the upstream repository, see the docs
for more details
https://freedesktop.pages.freedesktop.org/ci-templates/ci-fairy.html#checking-merge-requests

Output is both on stdout and as junit file which is presented nicely by
gitlab. Provided anyone reads it...

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>