tests: add a test to compile all xkb_symbols section with xkbcomp
Using pytest because it does a lot of the setup/tracing stuff for us. The test
checks all symbols files and generates a list of tuples (layout, variant)
from the files. That's then used in a generic-enough keymap to be fed to
xkbcomp. Where xkbcomp fails we fail the test (and pytest will collect
stdout/stderr/etc.) for us.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>