gitlab CI: run the xkeyboard-config rules test tool against evdev.extras.xml
libxkbcommon's parser has an optimization where it stops parsing a file when
the required keymap is found. The result of that was that the syntax error
fixed in !83 didn't get detected - all tested layouts where before that
erroneous line.

There's only one xkb_symbols map thats below that line but that one is
referenced in evdev.extras.xml, not evdev.xml. So let's update our CI to test
for that too.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>