rules: fold rules/compat/Makefile.am into rules/Makefile.am
Some files in rules/compat are generated but we need those files as input to
the various scripts in the rules directory. With separate Makefiles, this
caused an issue when running distcheck - some of the files were only available
after a make all, not with an immediate make distcheck.

Fix this by merging the two Makefiles, together with the various required
renames and general shuffling around.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>