rules: improve two layout descriptions, and drop useless short ones
Standardize the descriptions of two recently added Indian layouts:
use English, and use the words "phonetic" and "Inscript" as elsewhere,
and lowercase "enhanced" as it is not a name.

ShortDescriptions that are not in English or just repeat the name of
the layout are not useful, so they are dropped.

Also, use the correct language codes.