symbols/gn: say that the layout is meant for typing N'Ko, not French
Further, it is not specific to Guinea, as the script is also used
in neighboring countries.  But the layout is meant for an AZERTY
keyboard, so its name should mention that.

Apart from fixing name and code and comment, also translate the
hex codes from the "0x0100" to the "U" format.

(The layout is somewhat weird, with widely separated parentheses
and a 'braceright' but not a 'braceleft', but that is something
for another time.)