symbols/nl: make the official layout match the standard exactly
Commit d1411e5e added an uppercase Ssharp to the fourth level of
several layouts.  That is fine for most layouts, but not for an
official one that should match the standard exactly.

The standard Dutch layout does not have a fourth level, so remove
the uppercase Ssharp.  Also, replace 'bar' with 'brokenbar' on the
third level of the LSGT key, even though it is a useless character.

Furthermore, tweak the comments, replace a dead URL, and reorder
a line to match the order in the diagram given in:

ftp://ftp.software.ibm.com/software/globalization/keyboards/KBD143.pdf