delete two useless "Sun dead keys" variants for Portuguese
The only difference between these variants and the ones that they
are based on is that they duplicate the dead diaeresis from the
third level also on the fourth level, overwriting the dead abovering
(AD11) or the dead agonek (AE12) that were there.  These variants
are not useful additions but instead useless clutter.