delete a confused "Sun dead keys" variant for Latin American
This variant added two dead keys meant for the Spanish layout to
the Latin American layout, meaning that braceleft and bracketleft
are no longer available on the first and second level of the AC11
key, but instead braceleft has moved to the third level of AC11
and bracketleft has moved to the third level of AD11... while
braceright and bracketright continue to sit on the first two
levels of AC12.  This is silly.  No one can be using this, and
there are no keyboards with keys marked in this way.  Just delete
this botched variant.