delete a useless "Sun dead keys" variant for Belgium
All that the variant does (for five keys) is copy the symbol from
the third level also to the fourth level, overriding a (possibly)
useful symbol that was already there.  Furthermore, on AD11 it puts
a bracketleft on the fourth level, but does not put a bracketright
on the fourth level of AD12.  This is silly.  This is not improving
things but making them worse.  Just remove the variant.