delete Icelandic "Sun dead keys" variant that lacks essential symbols
The "Sun dead keys" variant for Icelandic adds the umlauted letters for
German (Ä, Ö, Ü) to the layout.  But Ä and Ü are not used in Icelandic,
and Ö is already there, on AE11, now duplicated on AC10.  Furthermore,
it *removes* the ETH and AE letters that are needed in Icelandic from
the lower levels, moves the often-used (on 5.8 percent of the letters)
dead acute accent from the comfortable AC11 to the uncomfortable AE12,
and also *removes* the minus, underscore, apostrophe, and question mark!

A layout where one cannot ask questions, cannot hyphenate words, cannot
apostrophize, and cannot type the essential letter ð (and æ only on the
third level), is useless.  Simply delete this botched variant.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>