delete the Icelandic 'nodeadkeys' variant, as it lacks essential symbols
With this variant, four symbols (ð - _ ?) are untypable.  It does have
an apostrophe (unlike the deleted "Sun dead keys" variant), but that
hardly makes this layout any more usable.  Just scrap it.

(Furthermore, the comment says "to not have any dead keys", but the
layout keeps the 'dead_belowdot' on level 3 of AB10, even though it
replaces 'dead_abovedot' on level 4 with a plain 'abovedot'.  When
keeping the 'dead_belowdot', why not also keep the 'dead_abovedot'?
Especially since there are no compose sequences with 'abovedot' --
the symbol cannot be used for anything but itself.  It has no use.)

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>