symbols/in: do not override essential digits with the Rupee sign
Commit 45521d61 from seven years ago indiscriminately added the
Rupee sign (₹) to the third level of many of the Indian layouts.
For most layouts this was not a problem, because they had nothing
on the third level of AE04.  But for four layouts this overwrote
the digit 4 or ४ or ൪ that already was there, thus making those
layouts inconsistent: having all the digits except "4".

Strangely, that same commit removed the dollar symbol that was
on the fourth level of AE04, which makes it seem that the intent
was to add the Rupee sign to the fourth level, not the third.
(See https://bugs.freedesktop.org/59157 for the actual rationale.)

Anyway...  For the four affected layouts, push the Rupee sign
to the fourth level of the "4" key.  People who *do* want the
sign on the third level can activate the relevant option.

Fixes #48.

Reported-by: Satyajeet Kanetkar