gitlab CI: check for new XF86 keysyms in the xorgproto repo
xorgproto 2021.2 and later has a recognizable pattern for adding new keysyms
to XF86keysym.h based on the Linux kernel input-event-codes.h. Use this to
detect any keysyms that are present in the header file but not yet in
symbols/inet.

This is merely a gitlab CI job as we only have to do an actual update once
every few months or so. A git diff is sufficient here too, it contains all the
information we will need to understand what is missing from the updated file.

We check xorgproto master because that gives us some heads-up on what will
come. There is a minor chance that we are mapping keycodes that change before
the release but it's minor and fixable anyway.

Requires: https://gitlab.freedesktop.org/xorg/proto/xorgproto/-/merge_requests/23

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>