Revert "gitlab CI: generate the evdev keycodes"
Some of the generated keys overwrote existing keys, causing warnings
and nonfunctional keys. For example:
xkbcommon: WARNING: Multiple names for keycode 121; Using <I121>, ignoring <MUTE>

Revert this commit, we're too close to a release and it's better to wait until
the next one to give this approach more time to settle.

Fixes #247

This reverts commit 5dc9b48c9b31de9f9780887a79ded3b1e52591d9.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>