keycodes: alias some used-by-default evdev keys
OUTP has been mapped to XF86Display in symbols/pc for about 13 years now, but
no evdev keycode is mapped to this alias. Instead, we have I235 mapped to
XF86Display in symbols/inet. Alias I235 to OUTP and remove the separate
mapping, the keymap stays the same.

Fixes xkbcomp warning
  Warning: Key <OUTP> not found in evdev+aliases(qwerty) keycodes
           Symbols ignored

The same goes for KITG/KIDN/KIUP, all of which are mapped in symbols/pc.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>