Restore "eu" Multi key
During the upgrade to v1.2 in 435d5612, Level 3 "m" was changed from
"Multi_key" to "dead_greek", following the upstream changes.
However, in the process Level 4 "m" was changed from "Multi_key" to "√".

As all variations of v1.2 and later available upsteam have some form of
compose key at L4 "m" and the .deb file available still it as "Multi_key",
this seems to have been a mistake.

It most likely happened because "√" is the placeholder symbol used in a
diagram on the upstream website to indicate the presence of a compose key
at that position.

Fixes #246