gitlab CI: route stderr/stdout into two different files
The tool prints to stderr for failures, stdout for success. So far we only
collected stdout but not stderr. For undefined keysyms we usually get lots of
failed keymaps and the log exceeds the gitlab limits.

Let's collect stdout and stderr as separate files instead, and save it as
artifacts on failure. We don't really care about the keymaps on success anyway
- no-one will look through gigabytes of keymaps in the hope of finding
something wrong.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>