rules: add a "custom" layout to the XML file
This layout does not exist and we will never provide it.

However, having it in the XML file means it will show up in GUI
configuration mechansism that parse the XML file directly (instead of using
libxbkcommon's libxkbregistry).

Our rulesets fall back to the file "symbols/layout", section "variant"
for any layout(variant) that's not explicitly covered. This enables users to
create a symbols/custom file with their layout and have it
available.

As there are no variants, the GUI tools will only be able to use the default
section. Commandline tools can use variants as well.

This is papering over the whole issue only, but it does provide for some
convenience. It will still require adding a file in /usr/share in most cases,
but since we do not provide this file, it will be safe from being overwritten.

Fixes #257

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>