symbols/us: May LSGT to backslash/bar on the us(euro) variant
In Europe some users may prefer to use an English (US) keyboard layout
over their country's native layout, there is enough demand for this that
keyboards with the English (US) layout are sold all over Europe.
In some countries (e.g. The Netherlands) the English (US) layout is even
the default layout.

The English (US) keyboards sold in Europe will have a euro-sign on the 5
key and in most cases they use the ISO physical-keyboard-layout which has
an extra 102nd key between the left Shift and the Z key. This extra key
is marked as a second backslash/bar key.

Before this commit the `us(euro)` variant did not contain a specific
binding for the `LSGT` key, so it inherited the standard less/greater
binding from the pc105 symbols, resulting in the key sending the wrong
symbols.

Add a `LSGT` key binding to backslash/bar to the `us(euro)` variant to
fix this.

Note under Windows even the `us(basic)` variant has the `LSGT` key bound
to backslash/bar, but a lot of XKB layouts inherit from the `us(basic)`
variant so changing it there will cause problems elsewhere.

Link: https://gitlab.freedesktop.org/xkeyboard-config/xkeyboard-config/-/merge_requests/198
Signed-off-by: Hans de Goede <hdegoede@redhat.com>