gitlab CI: parse through the YAML file to list failed keymap compilations
libxkbcommon (commit 1cae25005211) now provides the output of the layout
tester format in YAML, with successful compilations on stdout and failed ones
on stderr. This makes it easy to collect the results, extract and print the
failures with yq but also parse the yaml file and leave a JUnit XML in place
that will then show up as result on a MR page.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>