SHIFT+base key on Keypad not modify the base key on keypad swaping to number key.
SHIFT+Page_Up, SHIFT+Home, etc are a very useful sequence keyboards in KDE, so
when I press SHIFT+Page_Up and when Page_Up is on keypad, gives number 9 and is
a very disapointing. For example in konsole, I use a lot SHIFT+Page_Up and
SHIFT+Page_Down to scroll the screen up and down.
After reasearch a lot if just not swap from base to number when press SHIFT,
while SHIFT+number is not problem send the base key solves the problem in a
very simple way

I found a similar request on this topic https://bbs.archlinux.org/viewtopic.php?id=257020
from where I based myself to find the solution.